Move UniqueValue para pacote 'compartilhado'
